chore: Hide internal types for "boringssl" backend
This commit hides the `BoringSha1` and `BoringSha256` types from the
`boringssl` backend. They technically have to be public because the
associated type of a public trait must be public, but they only exist to
wrap the types from the `boring` crate and implement the relevant traits
from the `digest` crate so those types can fit into the interface we
require for a cryptography backend. We don't _really_ want downstream
users to use these types, so we make them public but `#[doc(hidden)]`.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>